refactor: address review-bot feedback on image themes
- Image cache: remove per-focus file-system I/O (cache hits are FS-free),
  bound the cache, and drop the modification-time stat on the render hot
  path (greptile P1 blocking-I/O). Document the small cache lock as the
  platform-bridge exception since the nonisolated AppKit apply path can't be
  actor-isolated without going async.
- Make applyImageThemePreset/clearBackgroundImageTheme async and run the
  disk writes off the main actor via Task.detached; decode the bundled asset
  to JPEG on the main actor first (CodeRabbit).
- Surface theme apply/clear failures via thrown ThemeError instead of
  best-effort logging.
- Move ImageThemePresets out of cmuxApp.swift into HostSettingsActions.swift;
  move ImageThemePresetInfo into its own package file with full DocC.
- Add DocC to the new public catalog keys.
- Remove the unreachable theme-frame reattach guard (dead code).
